git - schema: Add new schema for adding git information
This CR adds a new ecs schema definitions for adding git related
information.  This will be useful for tagging messages with the
specific git repository and commit hash related to the code which
produced the message.

Specifically I would like to add the git hash of the puppet repo
responsible for generating a specific puppet report[1]

[1]https://gerrit.wikimedia.org/r/c/operations/puppet/+/719368

This CR also adds vim tmp files to gitignore

Bug: T222826
Change-Id: Ib7f1607497df6d006bf67c91976930e825f37823