validator: fix get_type to correctly handle ints
JS is only aware of a Number type stored in IEEE 754 double, therefore
every Number is a float.  Because of this, toString(float) is
insufficient to detect if the value supplied is a true float, but will
detect if the right side of the decimal point is greater than zero.
To work around this, floating point data types will not indicate a
problem when encountering an integer.

Bug: T234565
Change-Id: Ieb23b214e3034225130efefbb558a8083ba8c440